# common/mempool.cc: Improve performance of sharding