--- a/src/test/test_mempool.cc
+++ b/src/test/test_mempool.cc
@@ -406,7 +406,7 @@ TEST(mempool, btree_map_test)
 
 TEST(mempool, check_shard_select)
 {
-  const size_t samples = mempool::num_shards * 100;
+  const size_t samples = mempool::num_shards * 30;
   std::atomic_int shards[mempool::num_shards] = {0};
   std::vector<std::thread> workers;
   for (size_t i = 0; i < samples; i++) {
@@ -421,16 +421,17 @@ TEST(mempool, check_shard_select)
   }
   workers.clear();
 
+#if !defined(MEMPOOL_SCHED_GETCPU)
   size_t missed = 0;
   for (size_t i = 0; i < mempool::num_shards; i++) {
-    if (shards[i] == 0) {
+    if (shards[i] != 30) {
+      // Each shard is expected to have exactly 30 threads
       missed++;
     }
   }
-
-  // If more than half of the shards did not get anything,
-  // the distribution is bad enough to deserve a failure.
-  EXPECT_LT(missed, mempool::num_shards / 2);
+  EXPECT_EQ(missed, 0u);
+#endif
+  // Else: test_c2c.cc is a better test of the sharding algorithm
 }