## Checklist
- Tracker (select at least one)
  - [x] References tracker ticket
  - [ ] Very recent bug; references commit where it was introduced
  - [x] New feature (ticket optional)
  - [ ] Doc update (no ticket needed)
  - [ ] Code cleanup (no ticket needed)
- Component impact
  - [ ] Affects [Dashboard](https://tracker.ceph.com/projects/dashboard/issues/new), opened tracker ticket
  - [ ] Affects [Orchestrator](https://tracker.ceph.com/projects/orchestrator/issues/new), opened tracker ticket
  - [x] No impact that needs to be tracked
- Documentation (select at least one)
  - [ ] Updates relevant documentation
  - [x] No doc update is appropriate
- Tests (select at least one)
  - [x] Includes [unit test(s)](https://docs.ceph.com/en/latest/dev/developer_guide/tests-unit-tests/)
  - [ ] Includes [integration test(s)](https://docs.ceph.com/en/latest/dev/developer_guide/testing_integration_tests/)
  - [ ] Includes bug reproducer
  - [ ] No tests

This is my 1st pull request and perhaps a little ambitious, so be gentle :-)

Improve the performance of memcache sharding by using sched_getcpu() to select which shard to use so that normally each shard is only accessed by a single CPU core to avoid CPU cache lines being shared between cores. Further improve performance by caching the shard selection in a thread local variable to mitigate the overheads of repeatedly calling sched_getcpu(). For builds that do not set thread affinity to a single core it is possible that kernel scheduler migrates the thread to a different core invalidating the cached result. Periodically check the cached value and update if necessary. Adapt how frequently this check is performed based on whether the thread was migrated with the aim to minimize the number of calls to sched_getcpu() and the percentage of the time that the cached result is invalid.

Additionally:

* Increase the number of shards on systems with more CPU cores, limit to a maximum of 128 to avoid non-performance crticial O(num shard) algorithms becoming too expensive
* For systems that do not support sched_getcpu() replace the current hash of pthread_self() with a round robin assignment of threads to shards. This fixes an arch64 issuehttps://tracker.ceph.com/issues/64100 where the current hashing ineffectively spread 3200 threads across just 4 of 32 available shards.

## Performance benchmarks

Lets start with some benchmarks to show the various costs of fragments of code. The objective is to show that the overheads of performing atomic accesses to a cache line that is dirty in another CPU core is far worse than the overheads of calling sched_getcpu():

These tests were run on a dual socket Intel Broadwell x86_64 system with no attempt to optimize the code being generated by the compiler, timings will vary on different systems and different architectures but it gives an idea of relative costs. The results are averages of many iterations and represent the cost when code+data is in the CPU L1 cache, in practice costs are likely to be higher because of CPU cache misses.

Calling sched_getcpu() to find the current CPU			32 cycles
Calling pthread_self() and hashing the result			8 cycles
Calling a function to return a thread_local int (local-exec)	5 cycles
Calling a function to return a thread_local int (global-dynamic) 15 cycles
Calling a function to return rdpid (assembly instruction for newer x86 systems) 5 cycles

Update: The thread_local int test was performed using a micro benchmark in a single executable which uses TLS local-exec which is the fastest way to access a thread local variable. It was pointed out that this is perhaps not representative and that we should really be comparing to TLS global-dynamic which is the worst case which happens when the thread local variable is in a dynamic library. I've added this extra benchmark which indeed shows some extra overheads.

Update: Investigated using rdpid as a faster alternative to sched_getcpu(). It appears that the rdpid instruction executes in 1 cycle on Intel Icelake CPUs, in this benchmark with the additional overheads of a function call it is being measured as 5 cycles. This is significantly faster than calling sched_getcpu() even though this is only reading a variable from the vDSO area on x86 systems.

Atomic increment of a memory variable				19 cycles
Non-atomic increment of a memory variable			5 cycles

Atomic increment with 8 threads (different cores, two sockets) accessing the same cache line in series			142 cycles intra socket, 400 cycles inter socket

This was a test with multiple threads performing atomic increments with the same cache line, but with serialization to ensure each thread did an increment in turn hence forcing the cache line to move between CPU cores each time, but with no contention. The cost of moving the cache line is higher when it needs to be moved between sockets.

Atomic increment with 2 threads (different cores of same socket) accessing the same cache line in parallel		75 cycles
Atomic increment with 4 threads (different cores of same socket) accessing the same cache line in parallel		166 cycles
Atomic increment with 6  threads (different cores of same socket) accessing the same cache line in parallel		253 cycles
Atomic increment with 8 threads (different cores of same socket) accessing the same cache line in parallel		330 cycles
Atomic increment with 8 threads (different cores, two sockets) accessing the same cache line in parallel			350 - 435 cycles

This was a test with multiple threads performing atomic increments within the same cache line to cause the cache line to ping pong between cores. The test ensured the threads all ran at the same time and generated contention where multiple threads are trying to lock the same cache line. The results suggest that threads were sometimes able to perform more than one increment before a cache line moved, but that when there is a lot of contention not only is there a cost of moving the cache line but also a cost of waiting for the cache line to become available

I also did some testing on a Mac M1 (only aarch64 system I’ve got access to), which shows that the costs are pretty similar. Mac does not have sched_getcpu() or methods for setting or testing thread affinity, so we need to assume that the kernel scheduler made sensible choices about where threads were run:

Calling pthread_self() and hashing the result			4.6ns
Calling a function to return a thread_local int			5.7ns

Atomic increment of memory variable				14.8ns
Non-atomic increment of memory variable			4.3ns
Atomic increment with 8 threads (same socket) accessing the same cache line in series		276ns
Atomic increment with 2 threads (same socket) accessing the same cache line in parallel		35ns
Atomic increment with 4 threads (same socket) accessing the same cache line in parallel		140ns
Atomic increment with 6  threads (same socket) accessing the same cache line in parallel		258ns
Atomic increment with 8 threads (same socket) accessing the same cache line in parallel		276ns

Conclusions

* Moving a CPU cache line between cores has a very high cost, much higher than calling sched_getcpu() to try and prevent this happening
* Caching the binding of a thread to a shard in a thread_local has a small performance benefit over repeatedly calling sched_getcpu()

## Comparison of different algorithms

There is an existing test case test_c2c which does a good job of measuring the performance of the mempool sharding code. It can be run with different numbers of threads and with/without sharding. I tried a variety of different algorithms and ran tests to compare the performance with different numbers of threads. The test system had 32 cores (dual socket, 8 cores + hyperthreading per socket). Tests were run with up to 128 threads to show behavior both when there are less threads than cores and when there are more threads than cores.

The X axis is number of threads, the Y access is number of operations completed in 30 seconds (higher is better):

![mempool](https://github.com/ceph/ceph/assets/156200352/85556643-655e-4325-9aca-f447504ed1ed)

* Without any sharding the cache line contention is terrible
* The pthread_self() hash works well on x86_64 until the number of threads exceeds the number of cores, then performance drops steeply
* A round robin assignment of threads to shards works as well as the pthread_self() hash but mitigates the problem with this hash on aarch64 platforms
* Calling sched_getcpu() every call to mempool gives more stable performance when the number of threads exceeds the number of cores, but the overheads of calling sched_getcpu() show that performance is worse than the existing pthread_self() hash for smaller numbers of threads
* The adaptive caching of sched_getcpu() gives the best performance, and in particular performance does not degrade as number of threads increases above the number of cores

This test case is perhaps a bit unrepresentative of real world performance because the threads are running without yielding and once all cores are busy there is little benefit in the kernel scheduler migrating threads.

## Additional statistics

To help evaluate how well the adaptive caching algorithm is working with real world tests additional statistics were added to the mempool dump. These statistics count how frequently the code is calling sched_getcpu() and how often it finds a thread is running on a different core to the cached value. Using the costs measured in the performance benchmarks above it is possible to calculate whether the adaptive caching is having a net performance benefit.

The statistics are output as part of the mempool dump: ceph daemon osd.NNN dump_mempools

correct_core is the number of times the cached value was for the correct core. This is approximate - it assumes that if a thread hasn't changed cores between two checks that the thread has remained running on the same core. It is possible the thread has switched cores multiple times ending up on the same core.

incorrect_core is the number of times the cached value was for the incorrect core. This is approximate - it assumes that if a thread has changed cores that on average 50% of the mempool updates made since the last check were on the wrong thread.

checked_core is the number of times the cached value was tested.

Percentage of mempool updates using the wrong shard = incorrect_core / (correct_core + incorrect_core) 
Percentage of mempool updates that called sched_getcpu = checked_core / (correct_core + incorrect_core)

The costs from the performance benchmarks above can be used to work out the efficiency of the caching algrotihm versus simply calling sched_getcpu every time:

cached_algorithm_cost = checked_core * sched_getcpu_cost + incorrect_core * cacheline_ping_pong_cost
uncached_algorithm_cost = (correct_core + incorrect_core) * sched_getcpu_cost

This can be used to show whether the cached algorithm is effective

## Tuning

The main part of the adaptive caching algorithm that can be tuned is how frequently to call sched_getcpu() to check the cached result.

If a check finds a thread has not moved the code increases the interval until the next test, if a thread has been migrated the code decreases the interval until the next test.

Intervals are measured in terms of number of calls to memcache by each thread.

Initially the code implemented an interval increase of 12.5% of the previous interval and an interval decrease of 75%. The cost of using an incorrect cached value is higher than calling sched_getcpu() so we want the decrease to be much more severe thanb the increase.

Testing running some I/O to a block volume showed that the algorithm was incorrect about 7% of the time and only checked the cached result 0.5% of the time. Calculating the costs showed that this still gave a significant saving versus calling sched_getcpu() every call.

Further testing refined the adaptive algorithm, with the delivered code implementing an increase of 1 and a decrease of 50%. This showed the algorithm was incorrect about 2.5% of the time and called sched_getcpu() about 2.5% of the time.

## Other observations

Because threads can migrate between shards there is no longer a guarantee that if a mempool allocation and free occur on the same thread that the counters for each shard will not go negative. However the only use of the shard data is to sum the values for all shards which will result in the correct value.

The use of size_t which is an unsigned value for performing the shard arithmetic gives the correct results but can cause arithmetic overflows. Some of the Crimson code uses ssize_t, however this is not really a signed type (theoretically only supports negative value -1) so I'm not convinced that it is any better.

<details>
<summary>Show available Jenkins commands</summary>

- `jenkins retest this please`
- `jenkins test classic perf`
- `jenkins test crimson perf`
- `jenkins test signed`
- `jenkins test make check`
- `jenkins test make check arm64`
- `jenkins test submodules`
- `jenkins test dashboard`
- `jenkins test dashboard cephadm`
- `jenkins test api`
- `jenkins test docs`
- `jenkins render docs`
- `jenkins test ceph-volume all`
- `jenkins test ceph-volume tox`
- `jenkins test windows`
- `jenkins test rook e2e`
</details>