@rzarzynski 2024-02-22
> Calling sched_getcpu() to find the current CPU 32 cycles
> Calling pthread_self() and hashing the result 8 cycles
> Calling a function to return a thread_local int 5 cycles

Are these numbers coming from a microbenchmark? Was `local-exec` be used as TLS model? I'm asking because things like `-fPIC` adds a degree of freedom.