@bill-scales 2024-02-23
> Are these numbers coming from a microbenchmark? Was `local-exec` be used as TLS model? I'm asking because things like `-fPIC` adds a degree of freedom.

These performance benchmarks were from a micro benchmark, timing millions of a calls to a function implementing the code being tested. I didn't even bother turning on optimization for the compiler, however the code and indeed the optimized sequence of uops is going to be L1 cached so this will make very little difference to the benchmarks.

The TLS model was local-exec, as you point out if the code is in a dynamic library the TLS access will be marginally slower.

However these are all second order effects. In practice this code isn't going to remain in the L1 cache so the costs of both a TLS access and a call to sched_getcpu will be higher than shown by this benchmark, but a TLS access is still access is still going to be something like a 1/3rd of the cost of calling sched_getcpu which will be something like 1/3rd to 1/8th of the cost of performing an atomic access to a dirty cache line in another core.