@aclamk 2024-02-27
@bill-scales An excellent measurement of cycle cost of different operations!

Only seeing the numbers I have an audacity to propose a significant rework of out mempool logic.

1) Drop `pick_a_shard` logic
2) Create `mempool_all` type that contains `bloom_filter` .. `unittest_2` counters (no __align__)
3) Make `mempool_all` a thread local ...
4) ... with constructor signaling to `mempool core` its presence
5) ... and destructor retrieving remaining values. (some values can be negative!)
6) Make `mempool core` that iterates active `mempool_all` tables for data retrieval:
    `allocated_bytes()`, `allocated_items()`

That will give us 5 or 15 cycles for TLS retrieval + 2 x 19 cycles for atomic increment, total of ~48 cycles per update.

Now for the things I am not certain:
7) Convert pool counter values from `atomic<size_t>` to regular `ssize_t`
8) Create signal-based (`man 2 tkill`) retrieval procedure that will circumvent the fact that counters are not atomic.

That will give us 5 or 15 cycles for TLS retrieval + 2 x 5 cycles for regular increment, total of ~20 cycles per update.

In this technique I expect to pay single penalty for core change. 
Rescheduling thread to different core requires a cache flush.
Thread resuming on different core will just experience stall at first access.

@bill-scales Do you think such rework can fly? Do you think it will be worthwhile?