@bill-scales 2024-02-28
>     1. Drop `pick_a_shard` logic
>     2. Create `mempool_all` type that contains `bloom_filter` .. `unittest_2` counters (no align to 128 bytes)
>     3. Make `mempool_all` a thread local ...
>     4. ... with constructor signaling to `mempool core` its presence
>     5. ... and destructor retrieving remaining values. (some values can be negative!)
>     6. Make `mempool core` that iterates active `mempool_all` tables for data retrieval:
>        `allocated_bytes()`, `allocated_items()`

I did consider something like this. If the thread count is relatively small then doing accounting per thread will be more efficient than per core and a technique like above will work well. However if there are 1000's of threads then accounting per thread will start to consume a lot of memory and trying to calculate summary counts across this number of shards will become expensive.

I don't have enough experience of Ceph to know how many threads is typical, I see between 30 and 70 threads in OSD processes (so this technique would work well there), I don't know how many threads (specifically those accessing mempool) is typical in clients.

I'm sure that the mempool stats/dump interfaces aren't called very often, so having these iterate across 1000 shards probably wouldn't be a problem, but there are other interfaces such as allocated_items and allocated_bytes that will be called more often, so number of shards is a concern.

Maybe others can comment on what typical thread counts might be and whether a shard per thread model is viable?

I did think about having a per thread sharding scheme and switching to a per core scheme if the number of threads got too high, however I thought that switching models without pausing threads is too complex.

> 7) Convert pool counter values from `atomic<size_t>` to regular `ssize_t`

If we did 1-6 then I think we would want to get rid of the atomic's as well. The challenge here is the memory consistency model of having one thread updating a value and another thread reading it. Not a problem for x86, and solvable for other architectures with a bit of care.

> 8) Create signal-based (`man 2 tkill`) retrieval procedure that will circumvent the fact that counters are not atomic.

This would be far too slow, my experience with tkill and signal handlers suggest its not uncommon for it to take 1ms for a thread to receive a signal from another thread (this was in an environment where the source and target thread had affinity to different cores and were the only processes running on those cores).

In an environment like Crimson you can dispatch a task on each I/O reactor to summarize per thread data, but we need something more generic